requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-ollama>=0.1.0
chromadb>=0.4.22
pypdf>=3.17.4
sentence-transformers>=2.3.1
python-dotenv>=1.0.0
requests>=2.31.0
